+++ plugins/export/pages/export.php
@@ -2,7 +2,7 @@
 
 $func = rex_request('func', 'string');
 $provider_id = (int) rex_request('provider_id', 'int');
-$used_machine_id = rex_request('used_machine_id', 'int');
+$used_machine_id = (int) rex_request('used_machine_id', 'int');
 
 /*
  * Do actions
